pyproject.toml
[project]
name = "trvel-fastapi"
version = "1.0.0"
description = "Trvel eSIM API - Premium travel eSIM provider for Australian travelers"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Trvel", email = "dev@trvel.co"}
]
keywords = ["esim", "travel", "fastapi", "api"]

dependencies = [
    # FastAPI & Web
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic[email]>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "httpx>=0.26.0",

    # Database
    "sqlalchemy>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pgvector>=0.2.4",

    # Redis
    "redis>=5.0.0",

    # Authentication
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",

    # Payments
    "stripe>=7.0.0",

    # eSIM
    "qrcode[pil]>=7.4.0",

    # LiveKit (Voice Agent)
    "livekit>=0.11.0",
    "livekit-agents>=0.8.0",

    # AI/ML
    "openai>=1.10.0",
    "anthropic>=0.18.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "tiktoken>=0.5.2",

    # Voice (STT/TTS)
    "deepgram-sdk>=3.0.0",
    "elevenlabs>=0.2.0",

    # Telephony
    "twilio>=8.10.0",

    # Utilities
    "tenacity>=8.2.0",
    "structlog>=24.1.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
